feat: 🎸 단위 계산함수 분리 및 곱하기 함수 추가
