+++ frontend/src/utils/index.ts
@@ -1,4 +1,4 @@
-export { parseNumberAndUnit } from './parseNumberAndUnit';
+export { parseNumberAndUnit, divideSize, multiplySize } from './calculateUtils';
 export { calculatePosition } from './arrangement';
 export { getRemainingSeconds } from './time';
 export { convertArrayToObject } from './participantUtils';
